feat(coordinator): finalize function in the scheduler module

+++ apps/coordinator/ts/scheduler/scheduler.service.ts
@@ -12,6 +12,7 @@ import type {
 import type { IIdentityScheduledPoll, IScheduledPoll } from "../redis/types";
 
 import { ErrorCodes } from "../common";
+import { ProofGeneratorService } from "../proof/proof.service";
 import { RedisService } from "../redis/redis.service";
 import { getPollKeyFromObject } from "../redis/utils";
 import { SessionKeysService } from "../sessionKeys/sessionKeys.service";
@@ -35,6 +36,7 @@ export class SchedulerService implements OnModuleInit {
     private readonly sessionKeysService: SessionKeysService,
     private readonly redisService: RedisService,
     private readonly schedulerRegistry: SchedulerRegistry,
+    private readonly proofService: ProofGeneratorService,
   ) {
     this.logger = new Logger(SchedulerService.name);
   }
@@ -64,9 +66,9 @@ export class SchedulerService implements OnModuleInit {
       return { isScheduled: false };
     }
 
-    try {
-      this.schedulerRegistry.getTimeout(key);
-    } catch (error) {
+    const doesExist = this.schedulerRegistry.doesExist("timeout", key);
+
+    if (!doesExist) {
       // delete redis data so user has to reschedule again
       await this.redisService.delete(key);
       return { isScheduled: false };
@@ -120,39 +122,85 @@ export class SchedulerService implements OnModuleInit {
   }
 
   /**
+   * Finalize poll starting from any step
+   *
+   * @param args - generate proofs arguments
+   */
+  async finalizePoll(poll: IScheduledPoll): Promise<void> {
+    const { maciAddress, pollId, chain, deploymentBlockNumber, mode, merged, proofsGenerated } = poll;
+
+    const key = getPollKeyFromObject({ maciAddress, pollId, chain });
+
+    try {
+      if (!merged) {
+        await this.proofService
+          .merge({
+            maciContractAddress: maciAddress,
+            pollId: Number(pollId),
+            chain,
+          })
+          .then(() => this.redisService.set(key, JSON.stringify({ ...poll, merged: true })))
+          .catch(async (error: Error) => {
+            await this.deleteScheduledPoll({ maciAddress, pollId, chain });
+            await this.setupPollFinalization(poll);
+
+            throw new Error(`Error merging poll ${pollId}: ${error.message}`);
+          });
+      }
+
+      if (!proofsGenerated) {
+        await this.proofService
+          .generate({
+            maciContractAddress: maciAddress,
+            poll: Number(pollId),
+            chain,
+            mode,
+            startBlock: deploymentBlockNumber,
+          })
+          .then(() => this.redisService.set(key, JSON.stringify({ ...poll, merged: true, proofsGenerated: true })))
+          .catch(async (error: Error) => {
+            await this.deleteScheduledPoll({ maciAddress, pollId, chain });
+            await this.setupPollFinalization({ ...poll, merged: true });
+
+            throw new Error(`Error generating proofs for poll ${pollId}: ${error.message}`);
+          });
+      }
+
+      await this.proofService
+        .submit({
+          maciContractAddress: maciAddress,
+          pollId: Number(pollId),
+          chain,
+        })
+        .then(() => this.deleteScheduledPoll({ maciAddress, pollId, chain }))
+        .catch(async (error: Error) => {
+          await this.deleteScheduledPoll({ maciAddress, pollId, chain });
+          await this.setupPollFinalization({ ...poll, merged: true, proofsGenerated: true });
+
+          throw new Error(`Error submitting proofs for poll ${pollId}: ${error.message}`);
+        });
+    } catch (error) {
+      this.logger.error(`Error finalizing poll ${pollId}: ${(error as Error).message}`);
+    }
+  }
+
+  /**
    * Store poll information and schedule its finalization.
    * @param args - Arguments for storing and scheduling poll finalization
    * @return delay in milliseconds until the poll is scheduled for finalization
    */
-  private async setupPollFinalization({
-    maciAddress,
-    pollId,
-    mode,
-    deploymentBlockNumber,
-    chain,
-    endDate,
-  }: IScheduledPoll): Promise<ISetupPollFinalizationResponse> {
+  private async setupPollFinalization(poll: IScheduledPoll): Promise<ISetupPollFinalizationResponse> {
+    const { maciAddress, pollId, chain, endDate } = poll;
     const key = getPollKeyFromObject({ maciAddress, pollId, chain });
 
-    await this.redisService.set(
-      key,
-      JSON.stringify({
-        maciAddress,
-        pollId,
-        mode,
-        deploymentBlockNumber,
-        chain,
-        endDate,
-      } as IScheduledPoll),
-    );
+    await this.redisService.set(key, JSON.stringify(poll));
 
     const shouldBeExecutedIn = endDate * 1000 - Date.now();
     const toBeExecutedIn = shouldBeExecutedIn > 0 ? shouldBeExecutedIn : 0;
     const delay = toBeExecutedIn + BUFFER_TIMEOUT;
 
-    const timeout = setTimeout(() => {
-      // TODO: execute this.finalizePoll (implementing this in the next PR)
-      this.redisService.delete(key);
+    const timeout = setTimeout(async () => {
+      await this.finalizePoll(poll);
     }, delay);
 
     const isPollTimeoutScheduled = this.schedulerRegistry.doesExist("timeout", key);
@@ -227,10 +275,9 @@ export class SchedulerService implements OnModuleInit {
   async deleteScheduledPoll(scheduledPoll: IIdentityScheduledPoll): Promise<IIsPollScheduledResponse> {
     const key = getPollKeyFromObject(scheduledPoll);
 
-    try {
+    const doesExist = this.schedulerRegistry.doesExist("timeout", key);
+    if (doesExist) {
       this.schedulerRegistry.deleteTimeout(key);
-    } catch (error) {
-      this.logger.error(`Failed to delete timeout for ${key}:`);
     }
 
     await this.redisService.delete(key);
@@ -272,16 +319,7 @@ export class SchedulerService implements OnModuleInit {
             return;
           }
 
-          const { delay } = await this.setupPollFinalization({
-            maciAddress: poll.maciAddress,
-            pollId: poll.pollId,
-            mode: poll.mode,
-            deploymentBlockNumber: poll.deploymentBlockNumber,
-            chain: poll.chain,
-            merged: poll.merged,
-            proofsGenerated: poll.proofsGenerated,
-            endDate,
-          });
+          const { delay } = await this.setupPollFinalization({ ...poll, endDate });
 
           this.logger.log(`poll ${value} to be executed at ${new Date(Date.now() + delay).toISOString()}`);
         } catch (error) {
